% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjam-package.R
\docType{package}
\name{tjam-package}
\alias{tjam}
\alias{tjam-package}
\title{tjam: Baixa e organiza dados processuais do TJAM}
\description{
Provê funções para baixar e organizar dados
    processuais do Tribunal de Justiça do Amazonas.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jjesusfilho/tjam}
  \item Report bugs at \url{https://github.com/jjesusfilho/tjam/issues}
}

}
\author{
\strong{Maintainer}: Jose de Jesus Filho \email{jose@consudata.com.br} (\href{https://orcid.org/0000-0001-7415-7383}{ORCID})

}
\keyword{internal}
