% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjam_baixar_merito.R
\name{tjam_baixar_merito}
\alias{tjam_baixar_merito}
\title{Baixa pdfs com decisões de mérito}
\usage{
tjam_baixar_merito(df, diretorio = ".")
}
\arguments{
\item{df}{Dataframe com processos e urls dos méritos}

\item{diretorio}{Diretório onde baixar os pdfs}
}
\value{
pdfs
}
\description{
Baixa pdfs com decisões de mérito
}
