% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjam_cpopg_baixar_tabela_docs.R
\name{tjam_cpopg_baixar_tabela_docs}
\alias{tjam_cpopg_baixar_tabela_docs}
\title{Baixar tabela docs}
\usage{
tjam_cpopg_baixar_tabela_docs(processos = NULL, diretorio = ".")
}
\arguments{
\item{processos}{processo}

\item{diretorio}{diretorio}
}
\value{
html/json
}
\description{
Baixar tabela docs
}
