% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjam_extrair_url_decisoes.R
\name{tjam_extrair_url_decisoes}
\alias{tjam_extrair_url_decisoes}
\title{Extrair urls das decis천es dos andamentos do cpopg}
\usage{
tjam_extrair_url_decisoes(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Informar arquivos}

\item{diretorio}{Informar diret처rio se n찾o informar arquivos}
}
\value{
tibble com processo, data_movimentacao, descricao e url.
}
\description{
Extrair urls das decis천es dos andamentos do cpopg
}
